Migrate test_get_next_results to Travis suite
