+++ setup.cfg
@@ -1,5 +1,5 @@
 [pycodestyle]
-exclude = .git,__pycache__,.vagrant,node_modules,tests/jenkins
+exclude = .git,__pycache__,.vagrant,node_modules
 # E121,E123,E126,E226,E24,E704,W503: Ignored in default pycodestyle config:
 # https://github.com/PyCQA/pycodestyle/blob/2.2.0/pycodestyle.py#L72
 # Our additions...
@@ -12,7 +12,7 @@ max-line-length = 140
 # flake8 is a combination of pyflakes & pycodestyle.
 # Unfortunately we have to mostly duplicate the above, since some tools use
 # pycodestyle's config (eg autopep8) so we can't just define everything under [flake8].
-exclude = .git,__pycache__,.vagrant,node_modules,tests/jenkins
+exclude = .git,__pycache__,.vagrant,node_modules
 # The ignore list for pycodestyle above, plus our own PyFlakes additions:
 # F403: 'from module import *' used; unable to detect undefined names
 # F405: name may be undefined, or defined from star imports: module
@@ -20,14 +20,14 @@ ignore = E121,E123,E126,E129,E226,E24,E704,W503,E501,F403,F405
 max-line-length = 140
 
 [isort]
-skip = .git,__pycache__,.vagrant,node_modules,migrations,tests/jenkins
+skip = .git,__pycache__,.vagrant,node_modules,migrations
 multi_line_output = 1
 force_grid_wrap = true
 line_length = 140
 
 [tool:pytest]
 testpaths = tests
-norecursedirs = __pycache__ jenkins ui
+norecursedirs = __pycache__ ui
 DJANGO_SETTINGS_MODULE=tests.settings
 # Disable unused auto-loaded plugins.
 addopts = --driver Firefox -p no:mozlog -p no:metadata -p no:html
