--- tests/jenkins/Jenkinsfile
@@ -1,68 +1,0 @@
-#!/usr/bin/env groovy
-
-/** Desired capabilities */
-def capabilities = [
-  browserName: 'Firefox',
-  version: '59.0',
-  platform: 'Windows 10'
-]
-
-pipeline {
-  agent any
-  libraries {
-    lib('fxtest@1.9')
-  }
-  options {
-    ansiColor('xterm')
-    timestamps()
-    timeout(time: 1, unit: 'HOURS')
-  }
-  environment {
-    PYTEST_PROCESSES = "${PYTEST_PROCESSES ?: "auto"}"
-    PYTEST_ADDOPTS =
-      "-n=${PYTEST_PROCESSES} " +
-      "--tb=short " +
-      "--color=yes " +
-      "--driver=SauceLabs " +
-      "--variables=capabilities.json"
-    PULSE = credentials('PULSE')
-    SAUCELABS = credentials('SAUCELABS')
-  }
-  stages {
-    stage('Test') {
-      steps {
-        writeCapabilities(capabilities, 'tests/jenkins/capabilities.json')
-        sh "tox -c tests/jenkins/tox.ini -e py27"
-      }
-      post {
-        always {
-          archiveArtifacts 'tests/jenkins/results/*'
-          junit 'tests/jenkins/results/*.xml'
-          submitToActiveData('tests/jenkins/results/py27_raw.txt')
-          submitToTreeherder('treeherder', 'e2e', 'End-to-end integration tests', 'tests/jenkins/results/*', 'tests/jenkins/results/py27_tbpl.txt')
-          publishHTML(target: [
-            allowMissing: false,
-            alwaysLinkToLastBuild: true,
-            keepAll: true,
-            reportDir: 'tests/jenkins/results',
-            reportFiles: "py27.html",
-            reportName: 'HTML Report'])
-        }
-      }
-    }
-  }
-  post {
-    failure {
-      emailext(
-        attachLog: true,
-        attachmentsPattern: 'tests/jenkins/results/py27.html',
-        body: '$BUILD_URL\n\n$FAILED_TESTS',
-        replyTo: '$DEFAULT_REPLYTO',
-        subject: '$DEFAULT_SUBJECT',
-        to: '$DEFAULT_RECIPIENTS')
-    }
-    changed {
-      ircNotification()
-    }
-  }
-}
