--- tests/jenkins/requirements.txt
@@ -1,7 +1,0 @@
-# pyup: ignore file
-mozlog==3.7
-PyPOM==1.3.0
-pytest==3.4.2
-pytest-metadata==1.6.0
-pytest-selenium==1.11.4
-pytest-xdist==1.22.2
