--- tests/jenkins/tox.ini
@@ -1,22 +1,0 @@
-[tox]
-skipsdist = true
-envlist = py27
-
-[testenv]
-passenv = BUILD_NUMBER DISPLAY GIT_BRANCH GIT_COMMIT GIT_URL JENKINS_URL \
-    JOB_NAME PYTEST_ADDOPTS PYTEST_BASE_URL SAUCELABS_USR \
-    SAUCELABS_PSW
-deps = -rrequirements.txt
-commands = pytest \
-    --junit-xml=results/{envname}.xml \
-    --html=results/{envname}.html --self-contained-html \
-    --log-raw=results/{envname}_raw.txt \
-    --log-tbpl=results/{envname}_tbpl.txt \
-    {posargs}
-
-[pytest]
-addopts = -n=auto --verbose -r=a --driver=Firefox
-testpaths = tests
-xfail_strict = true
-base_url = https://treeherder.allizom.org
-sensitive_url = mozilla\.org
